+++ test/sys/test_aio_drop.rs
@@ -1,18 +1,23 @@
 extern crate nix;
 extern crate tempfile;
 
-use nix::sys::aio::*;
-use nix::sys::signal::*;
-use std::os::unix::io::AsRawFd;
-use tempfile::tempfile;
-
 // Test dropping an AioCb that hasn't yet finished.
 // This must happen in its own process, because on OSX this test seems to hose
 // the AIO subsystem and causes subsequent tests to fail
 #[test]
 #[should_panic(expected = "Dropped an in-progress AioCb")]
-#[cfg(not(target_env = "musl"))]
+#[cfg(all(not(target_env = "musl"),
+          any(target_os = "linux",
+              target_os = "ios",
+              target_os = "macos",
+              target_os = "freebsd",
+              target_os = "netbsd")))]
 fn test_drop() {
+    use nix::sys::aio::*;
+    use nix::sys::signal::*;
+    use std::os::unix::io::AsRawFd;
+    use tempfile::tempfile;
+
     const WBUF: &[u8] = b"CDEF";
 
     let f = tempfile().unwrap();